+++ Backend/daw2026/src/main/java/daw2026/Dto/UpdateLocalRequest.java
@@ -1,5 +1,7 @@
 package daw2026.Dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,9 +12,9 @@ import lombok.NoArgsConstructor;
 public class UpdateLocalRequest {
     
     private String name;
-    private int latitude;
-    private int longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private String ubication;
-    private int capacity;
-    private int rooms;
+    private Integer capacity;
+    private Integer rooms;
 }
